Added Behavior_Score attribute to Ratings. Added edmx model changes to correspond with DB table change
Added overloaded matchmaker method that includes behavior score in its decision making. If a bool value is passed as a third parameter to the matchmaker method, it will order players by their behavior score in ascending order. The design assumption is that bad behavior will increase the behavior score and thus reduce a player's priority in the matchmaker.

Added unit tests for the behavior score functionality